code review :  rename entity GlobalQuestion by DatasetQuestion
